Fix more reader crash bugs found by AFL fuzzer
- DeltaBinaryPackedDecoder: add span-based ReadInt32/ReadInt64 overloads
  for zero-allocation decoding into pre-allocated buffers; add
  TryDecodeDeltaBinaryPackedIntoBuffer in ColumnChunkReader; replace
  checked((int)...) header reads with ReadHeaderVarInt helper that throws
  CorruptParquetException instead of OverflowException
- CompactProtocolReader: add ReadVarU32(max) for raw unsigned varints
  (binary lengths and list counts are NOT zigzag-encoded); fix Skip(Binary)
  and ReadListHeader to use ReadVarU32 instead of ReadI32AsU32; add
  ReadListHeader overflow fix; make Remaining return uint and remove casts
- ReaderAllocationTests: add allocation tests for DeltaBinaryPacked,
  DeltaByteArray, DeltaLengthByteArray (latter two kept failing as
  reminders to implement zero-copy byte[] decoding)

+++ Plank/Reading/PageHeaderReader.cs
@@ -26,10 +26,10 @@ static class PageHeaderReader
                     type = (PageHeaderType)reader.ReadI32();
                     break;
                 case 2:
-                    uncompressedPageSize = (int)reader.ReadI32AsU32(max: (uint)reader.Remaining);
+                    uncompressedPageSize = (int)reader.ReadI32AsU32(max: reader.Remaining);
                     break;
                 case 3:
-                    compressedPageSize = (int)reader.ReadI32AsU32(max: (uint)reader.Remaining);
+                    compressedPageSize = (int)reader.ReadI32AsU32(max: reader.Remaining);
                     break;
                 case 7:
                     valueCount = ReadDictionaryHeader(ref reader);
@@ -88,10 +88,10 @@ static class PageHeaderReader
                     encoding = ParquetMetadataThriftReader.ReadEncoding(reader.ReadI32());
                     break;
                 case 5:
-                    definitionLevelsByteLength = reader.ReadI32AsU32(max: (uint)reader.Remaining);
+                    definitionLevelsByteLength = reader.ReadI32AsU32(max: reader.Remaining);
                     break;
                 case 6:
-                    repetitionLevelsByteLength = reader.ReadI32AsU32(max: (uint)reader.Remaining);
+                    repetitionLevelsByteLength = reader.ReadI32AsU32(max: reader.Remaining);
                     break;
                 case 7:
                     isCompressed = reader.ReadBool(inlineBool);
